# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Backend API",
    "version": "1.0.0",
    "description": "API documentation cho hệ thống Learning Management System (LMS)",
    "contact": {
      "name": "LMS Team",
      "email": "admin@lms.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.lms.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token để xác thực"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "name",
          "email",
          "role"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID duy nhất của người dùng",
            "example": 1
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Tên người dùng",
            "example": "Nguyễn Văn A"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email của người dùng",
            "example": "nguyenvana@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "teacher",
              "student"
            ],
            "description": "Vai trò của người dùng",
            "example": "student"
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 20,
            "description": "Số điện thoại của người dùng",
            "example": "0901234567"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "description": "Mã số sinh viên hoặc mã số giáo viên (duy nhất)",
            "example": "SV001"
          },
          "address": {
            "type": "string",
            "description": "Địa chỉ của người dùng",
            "example": "Hà Nội"
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái kích hoạt tài khoản",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Thời gian tạo tài khoản",
            "example": "2023-10-07T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Thời gian cập nhật cuối",
            "example": "2023-10-07T10:30:00Z"
          }
        }
      },
      "UserRegister": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Tên người dùng",
            "example": "Nguyễn Văn A"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email của người dùng",
            "example": "nguyenvana@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)",
            "description": "Mật khẩu (ít nhất 6 ký tự, có chữ hoa, chữ thường và số)",
            "example": "Password123"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "teacher",
              "student"
            ],
            "description": "Vai trò của người dùng",
            "example": "student",
            "default": "student"
          }
        },
        "description": "Sau khi đăng ký, tài khoản sẽ có: code = null (chưa có mã số), isActive = false (chưa được kích hoạt). Admin sẽ thêm mã số và kích hoạt tài khoản sau."
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Thành công"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Có lỗi xảy ra"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Email không hợp lệ"
                }
              }
            }
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "totalUsers": {
            "type": "integer",
            "example": 50
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Lấy danh sách người dùng thành công"
          },
          "data": {
            "type": "object",
            "properties": {
              "users": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/PaginationInfo"
              }
            }
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "integer",
            "example": 100
          },
          "newUsersThisMonth": {
            "type": "integer",
            "example": 15
          },
          "roleStats": {
            "type": "object",
            "properties": {
              "admin": {
                "type": "integer",
                "example": 2
              },
              "teacher": {
                "type": "integer",
                "example": 10
              },
              "student": {
                "type": "integer",
                "example": 88
              }
            }
          }
        }
      },
...
    },
...
  },
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Lấy danh sách tất cả người dùng (Admin only)",
        "description": "Lấy danh sách người dùng với phân trang, tìm kiếm và lọc",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Số lượng user mỗi trang"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "teacher",
                "student"
              ]
            },
            "description": "Lọc theo vai trò"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Tìm kiếm theo tên hoặc email"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "email",
                "role"
              ],
              "default": "createdAt"
            },
            "description": "Sắp xếp theo trường"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "DESC"
            },
            "description": "Thứ tự sắp xếp"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập (chỉ Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Tạo người dùng mới (Admin only)",
        "description": "Admin tạo tài khoản người dùng mới",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo người dùng thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ hoặc email đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập (chỉ Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/stats": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Lấy thống kê người dùng (Admin only)",
        "description": "Lấy thống kê tổng quan về người dùng trong hệ thống",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thống kê thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thống kê người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserStats"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập (chỉ Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}/reset-password": {
      "put": {
        "tags": [
          "User Management"
        ],
        "summary": "Reset mật khẩu người dùng (Admin only)",
        "description": "Admin reset mật khẩu cho người dùng",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID của người dùng cần reset mật khẩu"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "newPassword"
                ],
                "properties": {
                  "newPassword": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 50,
                    "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)",
                    "description": "Mật khẩu mới",
                    "example": "NewPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset mật khẩu thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập (chỉ Admin)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}